Fix error in Event Hubs multiple event batch sample
The sample has an error in it that will result in a message being lost
each time a new batch is created. The original implementation will
remove an event from the queue and then try to add it to the batch. If
the event cannot be added to the batch, the event is discarded as the
loop proceeds to the next iteration.

This implementation uses the Peek method to test whether the event can
be added to the batch and only dequeues it from the queue if it is
successfully added to the batch.